Extract OnConflictSupport
Update README
